[BEBIG-139] refactor : delete param null값 처리

+++ src/main/java/beBig/controller/CommunityController.java
@@ -109,6 +109,9 @@ public class CommunityController {
 
     @DeleteMapping("/{postId}/delete")
     public ResponseEntity delete(@PathVariable Long postId) {
+        if(postId ==null){
+            throw new NoContentFoundException("No content found for the given filters.");
+        }
         log.info("delete community");
         communityService.delete(postId);
         return ResponseEntity.status(HttpStatus.OK).build();
